requirements.txt
PySide6
ftfy==6.1.1
ipython==8.12.3
numpy==1.24.3
opencv_python==4.10.0.84
pandas==2.0.3
Pillow==9.3.0
regex
scikit_learn==1.3.2
scipy==1.10.1
setuptools==59.5.0
scikit-image==0.21.0
tqdm==4.66.5
torch
torchvision
